Make website generation insensitive to buildcache layout
Do not assemble the url to the root buildcache for each ref, instead
take the url of the "root" element provided by the cache indexer, and
just replace the s3-protocol/bucket prefix with the cloudfront https
prefix.

Also remove the column with download tarball links from the specs
table, as those are not useful without the metadata, and should be
downloaded by installing them with spack anyway.